Add statsd support to Ambassador without socat
This commit fixes the broken functionality of exporting
metrics over statsd in Ambassador. Besides fixing, the new
implementation does not require `socat` binary to forward
the metrics, instead the metrics are directly sent to
STATSD_HOST by envoy, which is `statsd-sink` by default.

statsd is only tested in the CI or when you're not in the
DEV mode, because DEV mode uses teleproxy to send packets
from ambassador to the cluster which does not support UDP
yet. Also, the tests are written around Graphite's statsd
API which is deployed to the cluster.

Fix #841