Make secret config non mandatory in TLS contexts
This commit makes adding valid secret config in a TLS context
optional. So, when `redirect_cleartext_from` or other valid fields
are specified but the secret specified is invalid, the TLS context
will still go through. This also means that when default
`ambassador-certs` secret is implied by Ambassador, and if it does
not exist, the rest of TLS Context goes through.

Also added tests.

Fix #1104