Set a capability to allow binding to low ports
The cap_net_bind_service capability allows a process to bind to a low
port without root. While Kubernetes lets you allow a capability in the
pod spec, the effective capabilities are wiped when the root -> user
transition is made.

This change `setcap`'s the binary so that it can assert that capability. e.g.
```
 securityContext:                                                                                                                                  |~
   capabilities:                                                                                                                                   |~
     add:                                                                                                                                          |~
       - NET_BIND_SERVICE
``` (there is no `CAP_` at the beginning; K8s seems to add that)

If you don't want to bind to a low port, don't add the capability and
we're back to the previous behavior. If you want to bind to a low port,
add the capability and configure Ambassador appropriately.