Makefile: Fix the envoy-src rule
Do you see what's wrong with this recipe:

	@PS4=; set -x; \
	git init $@
	cd $@ && \
	if git remote get-url origin &>/dev/null; then \
		git remote set-url origin $(ENVOY_REPO); \
	else \
		git remote add origin $(ENVOY_REPO); \
	fi && \
	git fetch --tags origin && \
	$(if $(filter-out -,$(ENVOY_COMMIT)),git checkout $(ENVOY_COMMIT),if ! git rev-parse HEAD >/dev/null 2>&1; then git checkout origin/master; fi)

Assuming you're a human and not a parser, you probably didn't notice the
missing "&& \" at the end of the "git init" line.

Wrap the whole thing in `{ }` to make sure that this type of thing gets
noticed.

While we're at it, unset GIT_DIR and GIT_WORK_TREE before calling nested
git, so that `git bisect` works properly (this trick came from the APro
Makefile).