Add kat docker images targets to test target
Some recent changes require kat client and server images to be
tagged and pushed before tests are run. Since these images are
required to run tests, this causes tests failures if they're not
pushed - which they are not during `make test`. This commit adds
both the kat targets to `make test` to fix this.