Makefile: Clean up `protoc` rules, get it generating Envoy .pb.go files again
The motivation here is that I wanted to

    import "github.com/gogo/protobuf/gogoproto/gogo.proto";

in Host.proto.  To do this, I would have to add stuff to the --proto_path
and set up a bunch of mappings for gogofast.

Having had to solve that exact problem several times before for Envoy,
I've gone ahead and revived those parts of cxx/envoy.mk

While I was at it, I fixed it so that `make generate` behaves correctly if
$(CURDIR) != $(OSS_HOME).

NB: This upgrades protoc from 3.5.1 to 3.6.1, which is an upgrade that
had been made previously.  IDK why it got downgraded.