Change resource name to store if duplicate found
Currently, resources stored in Ambassador are stored as
{resource_name: resource object, ...} treating resource name as
a unique identifier to look up resources of a given kind. However,
this causes issues becuase this does not allow different resources
in different namespace to have the same name, hence resulting in
errors or ignoring these resources completely.

In this commit, if two (or more) resources have the same name (in
different namespaces), then the duplicate resources are stored
as {name.namespace: resource} instead of just {name: resource}.

An ideal solution would be to idenitfy all resources with
{
  name: resource_name,
  namespace: resource_namespace
}
instead of flat strings, but that comes later.