Introduce abstractions for K8s resource processing
This change adds a set of new configuration resource processing classes
that reduce the size of the resource fetcher and delegate most of the
work associated with handling a particular resource kind to
single-purpose processor objects.

Initially, we're only porting the Knative handling (specific to this
changeset), but the processors are designed to handle the other resource
types as well.

A management class, ResourceManager, glues everything together with
common context. Notably, this includes the set of final ACResource
elements to emit and location information as propagated by the fetcher.

We also add a few utility classes: ResourceDict, which provides some
helpers to access common information from a Kubernetes resource;
ResourceKind, a wrapper around a Kubernetes resource type's API group,
version, and kind (GVK); ResourceIdent, which lets us uniquely identify
a given resource and implements __hash__; and ResourceEmission, which
encapsulates the object emission and its corresponding rkey before final
processing into an ACResource.