(from AES) generate.mk: Hack around bug in controller-gen
When I ran `make update-yaml`, I was getting different results than CI.
The root cause is a bug in controller-gen:

  https://github.com/kubernetes-sigs/controller-tools/pull/485

However, the reason the behavior was different on my laptop than in CI is
that activating the bug depends on the order of the arguments to
controller-gen.  And the order of the arguments turns out to depend on
the order of $(.VARIABLES), which is ordered differently in different
versions of GNU Make.  And my laptop and CI have different versions of
Make.

So, add some determinism and wrap it in $(sort …) so that the bug at least
triggers consistently.